+++ src/helpers.js
@@ -15,3 +15,11 @@ export function makeCourseList(x){
     else
         return x;
 }
+
+export function createDicts (area) {
+    let arrayOfDict = []
+    for (let i = 0 ; i < area.length ; i++) {
+        arrayOfDict.push({category:area[i], selected: false});
+    }
+    return arrayOfDict;
+}
